Use double

+++ sqml-api/src/main/java/ai/dataeng/sqml/analyzer2/SqrlSinkBuilder.java
@@ -571,7 +571,7 @@ public class SqrlSinkBuilder {
 
     @Override
     public String visit(DoubleType doubleType) {
-      return "DOUBLE";
+      return "DECIMAL";
     }
 
     @Override
